CellCycleScoring_UCell

Similar to Seurat's CellCycleScoring, except using UCell AddModuleScore_UCell instead of Seurat's AddModuleScore

CellCycleScoring_UCell(
  seuratObj,
  outputFieldName = "Phase",
  min.genes = 10,
  s.genes = GetSPhaseGenes(),
  g2m.genes = GetG2MGenes(TRUE),
  assayName = "RNA",
  min.score.G2M = 0.02,
  min.score.S = 0.02,
  facetField = "ClusterNames_0.2",
  ncores = 1
)

Arguments

seuratObj

The seurat object

outputFieldName

The name of the meta.data field to store results

min.genes

If less than min.genes are shared between the seurat object and the reference cell cycle genes, this method will abort.

s.genes

A vector of features associated with S phase

g2m.genes

A vector of features associated with G2M phase

assayName

The name of the assay

min.score.G2M

The minimum UCell score needed to consider a cell positive for G2M

min.score.S

The minimum UCell score needed to consider a cell positive for S

facetField

If not NA, a plot will be produced summarizing S.Core and G2M.Score, faceted by this value.

ncores

Passed directly to AddModuleScore_UCell

Value

A modified Seurat object.