ClrNormalizeByGroup

This subsets the input object based on a variable (like dataset), and performs CLR normalization per-group, then combines them

ClrNormalizeByGroup(
  seuratObj,
  groupingVar,
  assayName = "ADT",
  targetAssayName = NA,
  margin = 1,
  minCellsPerGroup = 20,
  calculatePerFeatureUCell = FALSE,
  featureInclusionList = NULL,
  featureExclusionList = NULL,
  doAsinhTransform = FALSE
)

Arguments

seuratObj

The seurat object

groupingVar

The variable to use to partition the data

assayName

The name of the assay

targetAssayName

If provided, data will be saved to this assay, rather than modifying the source assay

margin

Passed directly to NormalizeData()

minCellsPerGroup

If provided, any group with newer than this many cells will be dropped

calculatePerFeatureUCell

If TRUE, UCell will be run once per feature in the assay

featureInclusionList

If provided, the input assay will be subset to just these features.

featureExclusionList

If provided, the input assay will be subset to exclude these features.

doAsinhTransform

If true, asinh transform will be performed on the raw counts prior to CLR