Wrapper around Seurat::CreateSeuratObject

Create Seurat object from count data (usually from Read10X()). This also sets pct.mito.

CreateSeuratObj(
  seuratData,
  datasetId,
  datasetName = NULL,
  minFeatures = 25,
  minCells = 0,
  mitoGenesPattern = "^MT-",
  annotateMitoFromReference = TRUE
)

Arguments

seuratData

Seurat input data, usually from Read10X().

datasetId

This will be used as a prefix for barcodes, and stored in metadata. Also used as the project name for the Seurat object.

datasetName

An optional print-friendly name that will be stored in metadata

minFeatures

Include cells where at least this many features are detected.

minCells

Include features detected in at least this many cells.

mitoGenesPattern

The expression to use when identifying mitochondrial genes

annotateMitoFromReference

If true, a list of mitochondrial genes, taken from (https://en.wikipedia.org/wiki/Category:Human_mitochondrial_genes) will be used to calculate p.mito

Value

A Seurat object with p.mito calculated.