CreateStudyWideBarPlot

Summarize the results of RunPairwiseContrasts() into a series of bar plots, showing the magnitude of differentially expressed genes over many contrasts.

CreateStudyWideBarPlot(
  pairwise_de_results,
  pvalue_threshold = 0.05,
  logFC_threshold = 1,
  LikeVsLike = F,
  LikeVsLikeGroupFieldIndex = NULL,
  facetLikeVsLike = T
)

Arguments

pairwise_de_results

The result of RunPairwiseContrasts().

pvalue_threshold

The pvalue/FDR threshold used to filter the differential expression tables for significance.

logFC_threshold

Log fold change threshold used to filter the differential expression tables for significance.

LikeVsLike

Boolean controlling whether or not one of the contrasts should be held constant in the summarization (i.e., only like cell-types should be considered, since Vaccinated_T cells vs Vaccianted_Myeloid cells will be very different, but not necessarily informative.)

LikeVsLikeGroupFieldIndex

The numerical index of the contrast_columns variable supplied to DesignModelMatrix() that should be held constant for like vs like comparisons.

facetLikeVsLike

Boolean controlling whether or not the final plot should be faceted in a LikeVsLike contrast. If you have several grouping variables, this could be helpful.

Value

A ggplot2 bar plot showing the magnitudes of the differentially expressed genes.