Find_Markers

Find_Markers

Find_Markers(
  seuratObj,
  identFields,
  outFile = NULL,
  testsToUse = c("wilcox", "MAST", "DESeq2"),
  numGenesToPrint = 20,
  onlyPos = F,
  pValThreshold = 0.001,
  foldChangeThreshold = 0.5,
  datasetName = NULL
)

Arguments

seuratObj
A seurat object

identFields
A vector of grouping fields for DE. Often these are the resolution, computed during FindClustersAndDimRedux()

outFile
A file where a table of markers will be saved

testsToUse
A vector of tests to be used. Each will be used to run FindAllMarkers() and the results merged. Available are: wilcox, bimod, roc, t, negbinom, poisson, LR, MAST, DESeq2

numGenesToPrint
The number of top markers per cluster to print in a table

onlyPos
If true, only positive markers will be saved

pValThreshold
Only genes with adjusted p-values below this will be reported

foldChangeThreshold
Only genes with log2 fold-change above this will be reported

datasetName
An optional label for this dataset. If provided, this will be appended to the resulting table.

Value

A DT::datatable object with the top markers, suitable for printing