Merge Seurat Objects

Merges a list of Seurat objects

MergeSeuratObjs(
  seuratObjs,
  projectName,
  merge.data = FALSE,
  expectedDefaultAssay = "RNA",
  excludedAssays = c("UCellRanks"),
  enforceUniqueCells = TRUE,
  errorOnBarcodeSuffix = FALSE,
  doGC = FALSE,
  duplicateBarcodeMode = "exclude-all"
)

Arguments

seuratObjs

A named list of seurat objects, optionally named (in which case these will be used as dataset names).

projectName

The project name when creating the final seurat object

merge.data

Passed directly to Seurat::merge

expectedDefaultAssay

If not null, the DefaultAssay on the resulting seurat object will be set to this

excludedAssays

An optional list of assay names to drop prior to merge.

enforceUniqueCells

If true, all inputs must have unique cellbarcodes.

errorOnBarcodeSuffix

In certain cases, software appends a digit (i.e. -1) to the end of cellbarcodes. These can be a problem when trying to make string comparisons. If true, the method will error if these are encountered.

doGC

If true, in an attempt to save memory gc() will be run after each seurat object is merged

duplicateBarcodeMode

This dictates the behavior when duplicate cell barcodes are encountered between merged objects. This can be either: 'exclude-all' (all duplicated dropped from all objects), or 'error'.

Value

A modified Seurat object.