PerformGlmFit

Creates an edgeR glm object

PerformGlmFit(
  seuratObj,
  design,
  test.use = "QLF",
  assayName = "RNA",
  minCountsPerGene = 1,
  legacy = FALSE,
  plotBCV = TRUE
)

Arguments

seuratObj

The seurat object

design

The model.matrix object

test.use

Can be either QLF or LRT. QLF runs edgeR::glmQLFTest, while LRT runs edgeR::glmLRT

assayName

The name of the assay to use

minCountsPerGene

Any genes with fewer than this many counts (across samples) will be dropped.

legacy

A passthrough variable for edgeR's glmQLF function. They recently (R 4.0) changed the default behavior, so this will break on earlier versions of R.

plotBCV

A boolean determining if the BCV plot should be shown.

Value

An edgeR glm object