PerformIntegration

This perform's Seurat's integration on a dataset

PerformIntegration(
  seuratObj,
  splitField = "SubjectId",
  nVariableFeatures = 4000,
  nIntegrationFeatures = 3500,
  k.weight = 20,
  dimsToUse = 20,
  integrationFeaturesInclusionList = NULL,
  integrationFeaturesExclusionList = NULL,
  minCellsPerSubsetObject = 75,
  dropGroupsBelowThreshold = TRUE
)

Arguments

seuratObj

A Seurat object.

splitField

The metadata field on which to split the seurat object for integration

nVariableFeatures

The number of variable features used for Seurat::FindVariableFeatures

nIntegrationFeatures

The number of features for Seurat::SelectIntegrationFeatures

k.weight

Passed to Seurat::IntegrateData()

dimsToUse

Passed to Seurat::IntegrateData(), as dims = 1:dimsToUse

integrationFeaturesInclusionList

An optional vector of genes that will be included in the integration genes

integrationFeaturesExclusionList

An optional vector of genes that will be excluded in the integration genes

minCellsPerSubsetObject

If any of the seurat objects after splitting on splitField have fewer than this many cells, they are either discarded or the function will throw an error, depending on dropGroupsBelowThreshold

dropGroupsBelowThreshold

If any of the seurat objects after splitting are below minCellsPerSubsetObject and if this param is true, they will be discarded.

Value

A modified Seurat object.