Run SingleR For A Seurat Object

Compute SingleR classification on a Seurat object

RunSingleR(
  seuratObj = NULL,
  datasets = c("hpca", "blueprint", "dice", "monaco"),
  assay = NULL,
  resultTableFile = NULL,
  rawDataFile = NULL,
  minFraction = 0.01,
  showHeatmap = TRUE,
  maxCellsForHeatmap = 20000
)

Arguments

seuratObj
A Seurat object

datasets
One or more datasets to use as a reference. Allowable values are: hpca, blueprint, dice, monaco, and immgen. See cellDex package for available datasets.

assay
The assay in the seurat object to use

resultTableFile
If provided, a table of results will be saved here

rawDataFile
If provided, the complete SingleR results will be saved to this file

minFraction
If provided, any labels present with fraction of this or fewer across cells will be converted to Unknown

showHeatmap
If true, heatmaps will be generated showing the SingleR calls

maxCellsForHeatmap
The heatmap will only be plotted if the total cells is below this number

Value

The modified seurat object