.GenerateSizeFactor

A helper function for ConstructEnrichmentDataFrameAndDoStatistics that calculates size factors for differently sized scRNA-Seq datasets.

.GenerateSizeFactor(
  seuratObj,
  normalizationField = "cDNA_ID",
  sizeFactorField = "SizeFactor",
  maxSizeFactor = 100
)

Arguments

seuratObj

The seurat object that holds the data.

normalizationField

The metadata column that is used for size factor calculation (equivalently, normalization between different datasets.) Recommended to be cDNA_ID.

sizeFactorField

The column name of the seruat object metadata that size factors should be stored in.

maxSizeFactor

The maximum allowable SizeFactor before an error is automatically thrown. For instance, a size factor of 100 means you have a group of 8000 cells that you're comparing to 80 cells.