Runs LDA Model

This function runs an LDA model on scRNA-seq expression data

runLDA(
  seuratObj,
  ntopics,
  alpha = 50,
  beta = 0.1,
  varFeatures = 5000,
  iterations = 500,
  burnin = 250,
  seed.number = GetSeed(),
  cores = 1,
  normalizationMethod = "CLR",
  assayName = Seurat::DefaultAssay(seuratObj),
  skipNormalization = FALSE
)

Arguments

seuratObj

Seurat object containing the data the model was created with.

ntopics

Number of topics to be used in the model. If parallel == TRUE, a vector of topics to run should be inputted

alpha

the value for alpha in the LDA model

beta

the value for beta in the LDA model

varFeatures

the number of variable features to use in the LDA model. The more features that are used, the slower the model will run and the more noise that will be introduced, but the model will be more complete in representing your entire dataset.

iterations

the number of iterations used when learning the LDA model.

burnin

number of iterations to run to allow the model to learn before calculating certain statistics. Models start at random points, so this allows model to get closer to the fit before certain statistics are calculated.

seed.number

random integer to set seed

cores

Number of cores to use, only applicable if parallel = TRUE

normalizationMethod

Normalization method used by Seurat NormalizeData. Options are CLR, LogNormalize and RC.

assayName

The name of the assay holding the source data

skipNormalization

If true, the data are assumed to be pre-normalized. Both normalization and Seurat::FindVarialeFeatures() are skipped. Therefore the arguments normalizationMethod and varFeatures are ignored.

Value

LDA Model

References

https://github.com/ohsu-cedar-comp-hub/TITAN

Author

TITAN