Find_Markers

Find_Markers

Find_Markers(
  seuratObj,
  identFields,
  outFile = NULL,
  testsToUse = c("wilcox", "MAST", "DESeq2"),
  numGenesToPrint = 20,
  onlyPos = F,
  pValThreshold = 0.001,
  foldChangeThreshold = 0.5,
  minPct = 0.1,
  minDiffPct = -Inf,
  datasetName = NULL,
  assayName = "RNA",
  verbose = FALSE,
  doPairwise = FALSE
)

Arguments

seuratObj

A seurat object

identFields

A vector of grouping fields for DE. Often these are the resolution, computed during FindClustersAndDimRedux()

outFile

A file where a table of markers will be saved

testsToUse

A vector of tests to be used. Each will be used to run FindAllMarkers() and the results merged. Available are: wilcox, bimod, roc, t, negbinom, poisson, LR, MAST, DESeq2

numGenesToPrint

The number of top markers per cluster to print in a table

onlyPos

If true, only positive markers will be saved

pValThreshold

Only genes with adjusted p-values below this will be reported

foldChangeThreshold

Only genes with log2 fold-change above this will be reported

minPct

Only test genes that are detected in a minimum fraction of min.pct cells in either of the two populations. Meant to speed up the function by not testing genes that are very infrequently expressed.

minDiffPct

Only test genes that show a minimum difference in the fraction of detection between the two groups.

datasetName

An optional label for this dataset. If provided, this will be appended to the resulting table.

assayName

The assay to use.

verbose

Passed to Seurat::FindMarkers

doPairwise

If true, rather than use Seurat::FindAllMarkers, the code will iterate each pair of values and generate a list of DEGs from these comparisons. This can help identify markers highly differentially expressed between some groups, but shared between others.

Value

A DT::datatable object with the top markers, suitable for printing